+++ addon/globalPlugins/TranslateAdvanced/app/utils/utils_network.py
@@ -11,6 +11,7 @@ import ctypes.wintypes
 import ssl
 import urllib.request
 import urllib.error
+import socket
 
 # Carga traducción
 addonHandler.initTranslation()
@@ -102,30 +103,17 @@ def realizar_solicitud_https(url):
 			raise
 ### Fin certificados
 
-def is_connected():
-	"""
-	Comprueba si hay conexión a internet utilizando ctypes.
-
-	:return: True si hay conexión, False en caso contrario.
-	"""
-	INTERNET_CONNECTION_OFFLINE = 0x20
-	INTERNET_CONNECTION_CONFIGURED = 0x40
-	INTERNET_CONNECTION_LAN = 0x02
-	INTERNET_CONNECTION_MODEM = 0x01
-	INTERNET_CONNECTION_PROXY = 0x04
-	INTERNET_RAS_INSTALLED = 0x10
-	INTERNET_CONNECTION_MODEM_BUSY = 0x08
-	INTERNET_CONNECTION_ONLINE = 0x01
-
-	# Cargar la biblioteca wininet.dll
-	wininet = ctypes.WinDLL("wininet.dll")
-
-	# Comprobar el estado de conexión
-	flags = ctypes.c_int(0)
-	connection = wininet.InternetGetConnectedState(ctypes.byref(flags), 0)
-
-	# Analizar los indicadores de conexión
-	if connection:
-		return not bool(flags.value & INTERNET_CONNECTION_OFFLINE)
-	else:
-		return False
+def check_internet_connection():
+    """
+    Comprueba si hay conexión a Internet intentando conectar a un host externo.
+    
+    Returns:
+        bool: True si hay conexión a Internet, False si no.
+    """
+    try:
+        # Intentar conectar a un host externo (8.8.8.8 es un servidor DNS público de Google)
+        socket.create_connection(("8.8.8.8", 53), timeout=5)
+        return True
+    except OSError:
+        pass
+    return False