+++ changelog.md
@@ -1,5 +1,6 @@
 * Agregado modulo de detector de idiomas.
 * Solucionado problema con la verbalización de algunos mensajes.
+* Cambiada forma de comprobar si hay internet.
 * Agregado retardo para iniciar el complemento si no se encuentra conexión a internet.
 *Corrección de errores
 - SHA256: